improve test

+++ Makefile
@@ -1,5 +1,6 @@
 init:
-		pip install -r requirements.txt
+		python -m pip install --upgrade pip
+		if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
 
 freeze:
 		rm requirements.txt
